Rename servlet to javax-servlet, add jakarta-servlet
