+++ gradle/libs.versions.toml
@@ -26,6 +26,8 @@ grails = "2.2.0"
 h2 = "1.3.173"
 hibernate = "4.3.5.Final"
 jackson = "2.10.3"
+jakarta-servlet-api = "6.0.0"
+jakarta-jsp-api = "3.1.0"
 javaee = "7.0"
 jooq-compile = "3.0.0"
 jooq-test = "3.9.5"
@@ -42,10 +44,12 @@ ratpack = "1.10.0-milestone-8"
 selenium-api = "3.0.0"
 selenium-runtime = "3.141.59"
 servlet-api = "2.3"
+javax-servlet-api = "4.0.1"
 shadow = "9.4.0"
 slf4j-api = "1.7.4"
 spock = "2.1-groovy-2.5"
-spring = "3.2.4.RELEASE"
+spring5 = "5.3.39"
+spring6 = "6.2.8"
 testcontainers = "2.0.3"
 
 [libraries]
@@ -62,6 +66,8 @@ groovy = { module = "org.codehaus.groovy:groovy-all", version.ref = "groovy" }
 h2 = { module = "com.h2database:h2", version.ref = "h2" }
 hibernate = { module = "org.hibernate:hibernate-core", version.ref = "hibernate" }
 jackson-databind = { module = "com.fasterxml.jackson.core:jackson-databind", version.ref = "jackson" }
+jakarta-servlet-api = { module = "jakarta.servlet:jakarta.servlet-api", version.ref = "jakarta-servlet-api" }
+jakarta-jsp-api = { module = "jakarta.servlet.jsp:jakarta.servlet.jsp-api", version.ref = "jakarta-jsp-api" }
 javaee = { module = "javax:javaee-api", version.ref = "javaee" }
 jooq-compile = { module = "org.jooq:jooq", version.ref = "jooq-compile" }
 jooq-test = { module = "org.jooq:jooq", version.ref = "jooq-test" }
@@ -76,6 +82,7 @@ selenium-api = { module = "org.seleniumhq.selenium:selenium-api", version.ref = 
 selenium-support = { module = "org.seleniumhq.selenium:selenium-support", version.ref = "selenium-runtime" }
 selenium-firefox-driver = { module = "org.seleniumhq.selenium:selenium-firefox-driver", version.ref = "selenium-runtime" }
 servlet-api = { module = "javax.servlet:servlet-api", version.ref = "servlet-api" }
+javax-servlet-api = { module = "javax.servlet:javax.servlet-api", version.ref = "javax-servlet-api" }
 ratpack-core = { module = "io.ratpack:ratpack-core", version.ref = "ratpack" }
 ratpack-groovy = { module = "io.ratpack:ratpack-groovy", version.ref = "ratpack" }
 ratpack-groovy-test = { module = "io.ratpack:ratpack-groovy-test", version.ref = "ratpack" }
@@ -85,7 +92,10 @@ ratpack-hikari = { module = "io.ratpack:ratpack-hikari", version.ref = "ratpack"
 ratpack-test = { module = "io.ratpack:ratpack-test", version.ref = "ratpack" }
 slf4j-api = { module = "org.slf4j:slf4j-api", version.ref = "slf4j-api" }
 spock = { module = "org.spockframework:spock-core", version.ref = "spock" }
-spring-test = { module = "org.springframework:spring-test", version.ref = "spring" }
+spring5-test = { module = "org.springframework:spring-test", version.ref = "spring5" }
+spring5-web = { module = "org.springframework:spring-web", version.ref = "spring5" }
+spring6-test = { module = "org.springframework:spring-test", version.ref = "spring6" }
+spring6-web = { module = "org.springframework:spring-web", version.ref = "spring6" }
 testcontainers-core   = { module = "org.testcontainers:testcontainers",   version.ref = "testcontainers" }
 testcontainers-junit5 = { module = "org.testcontainers:testcontainers-junit-jupiter", version.ref = "testcontainers" }
