+++ integration-test/glassfish4/src/main/webapp/WEB-INF/web.xml
@@ -23,7 +23,7 @@
 
 	<filter>
 		<filter-name>miniprofiler</filter-name>
-		<filter-class>io.jdev.miniprofiler.servlet.ProfilingFilter</filter-class>
+		<filter-class>io.jdev.miniprofiler.javax.servlet.ProfilingFilter</filter-class>
 		<!-- integration test for non-standard path too -->
 		<init-param>
 			<param-name>path</param-name>
