+++ integration-test/jakarta-servlet/src/integrationTest/resources/META-INF/services/org.junit.platform.launcher.LauncherSessionListener
@@ -1,0 +1,1 @@
+io.jdev.miniprofiler.integtest.JakartaJettyContainerManager
