+++ integration-test/jakarta-servlet/src/main/resources/miniprofiler.properties
