--- integration-test/servlet/src/integrationTest/groovy/io/jdev/miniprofiler/servlet/funtest/HomePage.groovy
@@ -1,27 +1,0 @@
-/*
- * Copyright 2013 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.jdev.miniprofiler.servlet.funtest
-
-import geb.Page
-import io.jdev.miniprofiler.test.pages.MiniProfilerModule
-
-class HomePage extends Page {
-    static url = ''
-    static content = {
-        miniProfiler(wait: true) { module MiniProfilerModule }
-    }
-}
