--- integration-test/servlet/src/integrationTest/groovy/io/jdev/miniprofiler/servlet/funtest/JspTagOverrideFunctionalSpec.groovy
@@ -1,33 +1,0 @@
-/*
- * Copyright 2018 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.jdev.miniprofiler.servlet.funtest
-
-import geb.spock.GebReportingSpec
-
-class JspTagOverrideFunctionalSpec extends GebReportingSpec {
-
-    void "can override properties on script tag"() {
-        when:
-        to HomePage, override: true
-
-        then: 'mini profiler visible with single timing info'
-        miniProfiler
-        miniProfiler.results.size() == 1
-        def result = miniProfiler.results[0]
-        result.button.time ==~ ~/\d+\.\d+ ms/
-    }
-}
