--- integration-test/servlet/src/main/webapp/WEB-INF/web.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" ?>
-<!--
-  ~ Copyright 2013 the original author or authors.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-
-<web-app xmlns="http://java.sun.com/xml/ns/javaee"
-		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-		 xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd"
-		 version="3.0">
-
-	<filter>
-		<filter-name>miniprofiler</filter-name>
-		<filter-class>io.jdev.miniprofiler.servlet.ProfilingFilter</filter-class>
-	</filter>
-
-	<filter-mapping>
-		<filter-name>miniprofiler</filter-name>
-		<url-pattern>/*</url-pattern>
-	</filter-mapping>
-
-	<listener>
-		<listener-class>io.jdev.miniprofiler.servlettest.Startup</listener-class>
-	</listener>
-</web-app>
