+++ miniprofiler-jakarta-servlet/src/main/resources/META-INF/miniprofiler.tld
@@ -1,0 +1,132 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright 2026 the original author or authors.
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~ http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<taglib xmlns="https://jakarta.ee/xml/ns/jakartaee"
+		xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+		xsi:schemaLocation="https://jakarta.ee/xml/ns/jakartaee https://jakarta.ee/xml/ns/jakartaee/web-jsptaglibrary_3_1.xsd"
+		version="3.1">
+
+	<description>Mini Profiler</description>
+	<tlib-version>1.0</tlib-version>
+	<short-name>miniprofiler</short-name>
+	<uri>https://miniprofiler-jvm.jdev.io/1.0</uri>
+
+	<tag>
+		<description>Writes a javascript tag that includes the Miniprofile Javascript user interface on the page.</description>
+		<name>script</name>
+		<tag-class>io.jdev.miniprofiler.jakarta.servlet.jsp.ScriptTag</tag-class>
+		<body-content>empty</body-content>
+
+		<attribute>
+			<description>Sets the profiler provider to use</description>
+			<name>profilerProvider</name>
+			<required>false</required>
+			<rtexprvalue>true</rtexprvalue>
+			<type>io.jdev.miniprofiler.ProfilerProvider</type>
+		</attribute>
+
+		<attribute>
+			<description>Sets the path to the UI resources</description>
+			<name>path</name>
+			<required>false</required>
+			<rtexprvalue>true</rtexprvalue>
+			<type>java.lang.String</type>
+		</attribute>
+
+		<attribute>
+			<description>Sets the position of the miniprofiler UI</description>
+			<name>position</name>
+			<required>false</required>
+			<rtexprvalue>true</rtexprvalue>
+			<type>java.lang.String</type>
+		</attribute>
+
+		<attribute>
+			<description>Sets the color scheme of the miniprofiler UI. Allowed values: Light, Dark, Auto.</description>
+			<name>colorScheme</name>
+			<required>false</required>
+			<rtexprvalue>true</rtexprvalue>
+			<type>java.lang.String</type>
+		</attribute>
+
+		<attribute>
+			<description>Sets the toggle shortcut for the UI</description>
+			<name>toggleShortcut</name>
+			<required>false</required>
+			<rtexprvalue>true</rtexprvalue>
+			<type>java.lang.String</type>
+		</attribute>
+
+		<attribute>
+			<description>Sets the maximum number of traces to show</description>
+			<name>maxTraces</name>
+			<required>false</required>
+			<rtexprvalue>true</rtexprvalue>
+			<type>java.lang.Integer</type>
+		</attribute>
+
+		<attribute>
+			<description>Sets the millisecond threshold for trivial timings</description>
+			<name>trivialMilliseconds</name>
+			<required>false</required>
+			<rtexprvalue>true</rtexprvalue>
+			<type>java.lang.Integer</type>
+		</attribute>
+
+		<attribute>
+			<description>Sets whether to show trivial timings by default</description>
+			<name>trivial</name>
+			<required>false</required>
+			<rtexprvalue>true</rtexprvalue>
+			<type>java.lang.Boolean</type>
+		</attribute>
+
+		<attribute>
+			<description>Sets whether to show child timings by default</description>
+			<name>children</name>
+			<required>false</required>
+			<rtexprvalue>true</rtexprvalue>
+			<type>java.lang.Boolean</type>
+		</attribute>
+
+		<attribute>
+			<description>Sets whether to show controls by default</description>
+			<name>controls</name>
+			<required>false</required>
+			<rtexprvalue>true</rtexprvalue>
+			<type>java.lang.Boolean</type>
+		</attribute>
+
+		<attribute>
+			<description>Sets whether the request is authorized view profiling results</description>
+			<name>authorized</name>
+			<required>false</required>
+			<rtexprvalue>true</rtexprvalue>
+			<type>java.lang.Boolean</type>
+		</attribute>
+
+		<attribute>
+			<description>Sets whether to hide the UI by default</description>
+			<name>startHidden</name>
+			<required>false</required>
+			<rtexprvalue>true</rtexprvalue>
+			<type>java.lang.Boolean</type>
+		</attribute>
+
+	</tag>
+
+</taglib>
