+++ settings.gradle.kts
@@ -14,7 +14,8 @@ includeBuild("gradle/plugins")
 
 include("miniprofiler-core")
 include("miniprofiler-javaee")
-include("miniprofiler-servlet")
+include("miniprofiler-javax-servlet")
+include("miniprofiler-jakarta-servlet")
 include("miniprofiler-test-support")
 include("miniprofiler-grails")
 include("miniprofiler-ratpack")
@@ -24,7 +25,8 @@ include("miniprofiler-viewer")
 include("doc:manual")
 
 include("integration-test:lib")
-include("integration-test:servlet")
+include("integration-test:javax-servlet")
+include("integration-test:jakarta-servlet")
 include("integration-test:glassfish4")
 include("integration-test:wildfly8")
 include("integration-test:ratpack")
